fix(editor): draw the editor preview with the shared renderer
